renamed app to postcodeinfo
