+++ .gitignore
@@ -1,5 +1,5 @@
 *.pyc
 .DS_Store
-addressfinder/local.py
+postcodeinfo/local.py
 env
 venv
