+++ docker/nginx.conf
@@ -77,8 +77,8 @@ server {
     gzip_proxied        any;
     gzip_types          text/plain text/xml application/xml application/xml+rss;
 
-   access_log      /var/log/nginx/addressfinder/access.log combined;
-   error_log       /var/log/nginx/addressfinder/error.log error;
+   access_log      /var/log/nginx/postcodeinfo/access.log combined;
+   error_log       /var/log/nginx/postcodeinfo/error.log error;
 
 #   auth_basic            "private";
 #   auth_basic_user_file  conf.d/htpassword;
@@ -92,7 +92,7 @@ server {
     }
 
     location /static/ {
-        alias /home/app/django/addressfinder/static/;
+        alias /home/app/django/postcodeinfo/static/;
     }
 }
 #mail {
