+++ docker/postcodeinfo.ini
@@ -9,8 +9,8 @@ master = true
 enable-threads = true
 processes = 2
 chdir = /home/app/django
-env = DJANGO_SETTINGS_MODULE=addressfinder.settings
-module = addressfinder.wsgi:application
+env = DJANGO_SETTINGS_MODULE=postcodeinfo.settings
+module = postcodeinfo.wsgi:application
 req-logger = file:/var/log/wsgi/request.log
 logger = file:/var/log/wsgi/error.log
 post-buffering = 1
