+++ postcodeinfo/wsgi.py
@@ -1,5 +1,5 @@
 """
-WSGI config for addressfinder project.
+WSGI config for postcodeinfo project.
 
 It exposes the WSGI callable as a module-level variable named ``application``.
 
@@ -10,7 +10,7 @@ https://docs.djangoproject.com/en/dev/howto/deployment/wsgi/
 import os
 from os.path import abspath, dirname
 from sys import path
-os.environ.setdefault("DJANGO_SETTINGS_MODULE", "addressfinder.settings")
+os.environ.setdefault("DJANGO_SETTINGS_MODULE", "postcodeinfo.settings")
 
 SITE_ROOT = dirname(dirname(abspath(__file__)))
 path.append(SITE_ROOT)
